Modify 'add & remove functions' in a single file
