+++ src/modules/__test__/add-remove.test.js
@@ -3,6 +3,9 @@
  */
 
 import { listen, addToLi } from '../addtask.mjs';
+import remove from '../remove.mjs';
+
+// Add Test
 
 describe('addToLi Function', () => {
   test('add Li element', () => {
@@ -29,5 +32,22 @@ describe('listen Function', () => {
 
     const list = document.querySelectorAll('#todoList li');
     expect(list).toHaveLength(1);
+  });
+});
+
+// Remove Test
+
+describe('remove Function', () => {
+  test('remove element', () => {
+    document.body.innerHTML = '<ul id="todoList"><li class="task"><input type="checkbox" name="0"><label for="0">Feed cats</label><div class="remove"><img src="img"></div></li></ul>';
+    const taskList = [{ name: 'Feed cats', index: 0, done: false }];
+    const item = document.querySelector('li');
+    const button = document.querySelector('.remove');
+
+    const todoList = document.querySelector('#todoList');
+    remove(item, 0, taskList, todoList);
+    button.click();
+    const list = document.querySelectorAll('#todoList li');
+    expect(list).toHaveLength(0);
   });
 });
